Refactor: Break down DiscoverStructure.ts into focused modules (#1395)
Extract types, static application methods, and Rust flush diagnostics
from the 5,413-line DiscoverStructure.ts into three focused modules:
- DiscoverStructureTypes.ts (297 lines): all shared interfaces
- DiscoveryApplication.ts (1,181 lines): creature modification methods
- RustFlushDiagnostics.ts (319 lines): flush validation and metrics

Main file reduced to 3,854 lines (~29% reduction) with full backward
compatibility maintained through re-exports. Includes 31 new tests.
